Catalog service with Java 21

+++ catalog-service/src/main/java/com/ecommerce/crtdev/catalog_service/domain/events/produces/ProductCreated.java
@@ -1,3 +1,5 @@
 package com.ecommerce.crtdev.catalog_service.domain.events.produces;
 
-public record ProductCreated (Long sellerId, String productId, String name, double price, int stock) {}
+import java.math.BigDecimal;
+
+public record ProductCreated (Long sellerId, String productId, String name, BigDecimal price, int stock) {}
