+++ catalog-service/src/main/java/com/ecommerce/crtdev/catalog_service/domain/events/produces/ProductPriceChanged.java
@@ -1,3 +1,5 @@
 package com.ecommerce.crtdev.catalog_service.domain.events.produces;
 
-public record ProductPriceChanged(String productId, double newPrice) {}
+import java.math.BigDecimal;
+
+public record ProductPriceChanged(String productId, BigDecimal newPrice) {}
